// packages/curve-ui-kit/src/themes/model/figma-tokens.ts
export const FIGMA_TOKENS = {
// ...
  surfacesAndText: {
    default: {
      light: {
        text: {
          primary: '#1f1e1bff',
          secondary: '#5a554fff',
          tertiary: '#968e84ff',
          highlight: '#3162f4ff',
          disabled: '#bbb6afff',
        },
        layer: {
          '1': {
            fill: '#f8f7f7ff',
            outline: '#d4d0ccff',
          },
          '2': {
            fill: '#eeecebff',
            outline: '#dedbd8ff',
          },
          '3': {
            fill: '#f8f7f7ff',
            outline: '#d4d0ccff',
          },
          highlight: '#3162f4ff',
        },
        tables: {
          'header fill': '#dedbd8ff',
        },
      },
// ...
    },
// ...
  },
// ...
} as const
export type FigmaTokens = typeof FIGMA_TOKENS